+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/asn1/Asn1Elements.kt
@@ -1,6 +1,7 @@
 package at.asitplus.signum.indispensable.asn1
 
 import at.asitplus.catching
+import at.asitplus.signum.indispensable.asn1.Asn1Element.Tag.Template.Companion.withClass
 import at.asitplus.signum.indispensable.io.ByteArrayBase64Serializer
 import io.matthewnelson.encoding.base16.Base16
 import io.matthewnelson.encoding.core.Decoder.Companion.decodeToByteArray
@@ -156,11 +157,11 @@ sealed class Asn1Element(
     fun asSet() = thisAs<Asn1Set>()
 
     /**
-     * Convenience function to cast this element to an [Asn1Tagged]
+     * Convenience function to cast this element to an [Asn1ExplicitlyTagged]
      * @throws Asn1StructuralException if this element is not an explicitly tagged structure
      */
     @Throws(Asn1StructuralException::class)
-    fun asTagged() = thisAs<Asn1Tagged>()
+    fun asExplicitlyTagged() = thisAs<Asn1ExplicitlyTagged>()
 
     /**
      * Convenience function to cast this element to an [Asn1EncapsulatingOctetString]
@@ -180,6 +181,53 @@ sealed class Asn1Element(
     private inline fun <reified T : Asn1Element> thisAs(): T =
         (this as? T)
             ?: throw Asn1StructuralException("${this::class.simpleName} cannot be reinterpreted as ${T::class.simpleName}.")
+
+
+    /**
+     * Creates a new implicitly tagged ASN.1 Element from this ASN.1 Element.
+     * NOTE: The [TagClass] of the provided [tag] will be used! If you want the result to have [TagClass.CONTEXT_SPECIFIC],
+     * use `element withImplicitTag (tag withClass TagClass.CONTEXT_SPECIFIC)`!. If a CONSTRUCTED Tag is applied to an ASN.1 Primitive,
+     * the CONSTRUCTED bit is overridden and set to zero.
+     */
+    inline infix fun withImplicitTag(tag: Tag): Asn1Element = when (this) {
+        is Asn1Structure -> {
+            if (tag.isConstructed) Asn1CustomStructure(
+                children,
+                tag.tagValue,
+                tag.tagClass,
+                sortChildren = isSorted
+            ) else Asn1CustomStructure.asPrimitive(children, tag.tagValue, tag.tagClass)
+        }
+        is Asn1Primitive -> Asn1Primitive(tag without CONSTRUCTED, content)
+    }
+
+    /**
+     * Creates a new implicitly tagged ASN.1 Element from this ASN.1 Element.
+     * Sets the class of the resulting structure to [TagClass.CONTEXT_SPECIFIC]
+     */
+    inline infix fun withImplicitTag(tagValue: ULong) = withImplicitTag(tagValue withClass TagClass.CONTEXT_SPECIFIC)
+
+
+    /**
+     * Creates a new implicitly tagged ASN.1 Element from this ASN.1 Structure.
+     * If the provided [template]'s tagClass is not set, the class of the resulting structure defaults to [TagClass.CONTEXT_SPECIFIC].
+     * If a CONSTRUCTED Tag is applied to an ASN.1 Primitive, the CONSTRUCTED bit is overridden and set to zero.
+     */
+    inline infix fun withImplicitTag(template: Tag.Template) = when (this) {
+        is Asn1Structure -> (this as Asn1Structure).withImplicitTag(
+            Tag(
+                tagValue = template.tagValue,
+                tagClass = template.tagClass ?: TagClass.CONTEXT_SPECIFIC,
+                constructed = template.constructed ?: tag.isConstructed
+            )
+        )
+
+        is Asn1Primitive -> Asn1Primitive(
+            Tag(template.tagValue, tagClass = template.tagClass ?: TagClass.CONTEXT_SPECIFIC, constructed = false),
+            content
+        )
+    }
+
 
     @Serializable
     @ConsistentCopyVisibility
@@ -217,7 +265,7 @@ sealed class Asn1Element(
             }
 
             val SET = Tag(tagValue = BERTags.SET.toULong(), constructed = true)
-            val ASN1_SEQUENCE = Tag(tagValue = BERTags.SEQUENCE.toULong(), constructed = true)
+            val SEQUENCE = Tag(tagValue = BERTags.SEQUENCE.toULong(), constructed = true)
 
             val NULL = Tag(tagValue = BERTags.ASN1_NULL.toULong(), constructed = false)
             val BOOL = Tag(tagValue = BERTags.BOOLEAN.toULong(), constructed = false)
@@ -239,7 +287,6 @@ sealed class Asn1Element(
             val TIME_GENERALIZED = Tag(tagValue = BERTags.GENERALIZED_TIME.toULong(), constructed = false)
             val TIME_UTC = Tag(tagValue = BERTags.UTC_TIME.toULong(), constructed = false)
 
-
         }
 
         val tagClass by lazy {
@@ -248,9 +295,9 @@ sealed class Asn1Element(
             }
         }
 
-        val isConstructed by lazy { encodedTag.first().toUByte().isConstructed() }
+        val isConstructed get() = encodedTag.first().toUByte().isConstructed()
 
-        val isExplicitlyTagged by lazy { isConstructed && tagClass == TagClass.CONTEXT_SPECIFIC }
+        internal val isExplicitlyTagged get() = isConstructed && tagClass == TagClass.CONTEXT_SPECIFIC
 
         override fun toString(): String =
             "${tagClass.let { if (it == TagClass.UNIVERSAL) "" else it.name + " " }}${tagValue}${if (isConstructed) " CONSTRUCTED" else ""}" +
@@ -287,6 +334,50 @@ sealed class Asn1Element(
         }
 
         override fun hashCode(): Int = encodedTag.contentHashCode()
+
+        /**
+         * creates a new Tag from this object, overriding the class. Useful for implicitTagging (see [Asn1Structure.withImplicitTag])
+         */
+        infix fun withClass(tagClass: TagClass) =
+            Tag(this.tagValue, tagClass = tagClass, constructed = this.isConstructed)
+
+        /**
+         * creates a new Tag from this object, negating the passed property. Useful for implicitTagging (see [Asn1Structure.withImplicitTag]).
+         * This is a NOOP for tag that don't have this bit set.
+         */
+        infix fun without(negated: TagProperty): Tag = when (negated) {
+            CONSTRUCTED -> Tag(this.tagValue, tagClass = this.tagClass, constructed = false)
+        }
+
+        /**
+         * A tag with optional tagClass and optional constructed indicator. Used for ASN.1 builder DSL
+         */
+        class Template(val tagValue: ULong, val tagClass: TagClass?, val constructed: Boolean?) {
+
+            /**
+             * Creates a new tag template from this template, negating the passed property
+             */
+            inline infix fun without(negated: TagProperty) = when (negated) {
+                is CONSTRUCTED -> Template(this.tagValue, this.tagClass, false)
+            }
+
+            companion object {
+                /**
+                 * Convenience function to construct a tag template from an ULong tag value and class
+                 */
+                inline infix fun ULong.withClass(tagClass: TagClass) =
+                    Template(tagValue = this, tagClass = tagClass, constructed = null)
+
+                /**
+                 * Convenience function to construct a tag from an ULong tag value and property
+                 */
+                inline infix fun ULong.without(negated: TagProperty) = when (negated) {
+                    is CONSTRUCTED -> Template(tagValue = this, tagClass = null, constructed = false)
+                }
+
+            }
+
+        }
     }
 }
 
@@ -306,8 +397,25 @@ object Asn1EncodableSerializer : KSerializer<Asn1Element> {
 /**
  * ASN.1 structure. Contains no data itself, but holds zero or more [children]
  */
-sealed class Asn1Structure(tag: Tag, children: List<Asn1Element>?) :
-    Asn1Element(TLV(tag, byteArrayOf()), children) {
+sealed class Asn1Structure(
+    /**
+     * The tag identifying this structure
+     */
+    tag: Tag,
+
+    /**
+     * This structure's child elements
+     */
+    children: List<Asn1Element>,
+    /**
+     * Whether this structure sorts child nodes or keeps them as-is.
+     * This **should** be true for SET and SET OF, but is set to false for SET and SET OF elements parsed
+     * from DER-encoded structures, because this has a chance of altering the structure for non-conforming DER-encoded
+     * structures.
+     */
+    val isSorted: Boolean = false
+) :
+    Asn1Element(TLV(tag, byteArrayOf()), if (!isSorted) children else children.sortedBy { it.tag }) {
 
 
     public override val children: List<Asn1Element>
@@ -337,12 +445,13 @@ sealed class Asn1Structure(tag: Tag, children: List<Asn1Element>?) :
      * Returns the current child (useful when iterating over this structure's children)
      */
     fun peek() = if (!hasMoreChildren()) null else children[index]
+
 }
 
 /**
  * Explicit ASN.1 Tag. Can contain any number of [children]
  */
-class Asn1Tagged
+class Asn1ExplicitlyTagged
 /**
  * @param tag the ASN.1 Tag to be used will be properly encoded to have [BERTags.CONSTRUCTED] and
  * [BERTags.CONTEXT_SPECIFIC] bits set)
@@ -351,6 +460,37 @@ class Asn1Tagged
  */
 internal constructor(tag: ULong, children: List<Asn1Element>) :
     Asn1Structure(Tag(tag, constructed = true, tagClass = TagClass.CONTEXT_SPECIFIC), children) {
+
+
+    /**
+     * Returns this [Asn1ExplicitlyTagged] children, if its tag matches [tag]
+     *
+     * @throws Asn1TagMismatchException if the tag does not match
+     */
+    @Throws(Asn1TagMismatchException::class)
+    fun verifyTag(explicitTag: Tag): List<Asn1Element> {
+        if (this.tag != explicitTag) throw Asn1TagMismatchException(explicitTag, this.tag)
+        return this.children
+    }
+
+    /**
+     * Returns this [Asn1ExplicitlyTagged] children, if its tag matches [tagNumber]
+     *
+     * @throws Asn1TagMismatchException if the tag does not match
+     */
+    @Throws(Asn1TagMismatchException::class)
+    fun verifyTag(tagNumber: ULong): List<Asn1Element> = verifyTag(Asn1.ExplicitTag(tagNumber))
+
+    /**
+     * Exception-free version of [verifyTag]
+     */
+    fun verifyTagOrNull(tagNumber: ULong) = catching { verifyTag(tagNumber) }.getOrNull()
+
+    /**
+     * Exception-free version of [verifyTag]
+     */
+    fun verifyTagOrNull(explicitTag: Tag) = catching { verifyTag(explicitTag) }.getOrNull()
+
     override fun toString() = "Tagged" + super.toString()
     override fun prettyPrint(indent: Int) = (" " * indent) + "Tagged" + super.prettyPrint(indent + 2)
 }
@@ -360,7 +500,7 @@ internal constructor(tag: ULong, children: List<Asn1Element>) :
  * @param children the elements to put into this sequence
  */
 class Asn1Sequence internal constructor(children: List<Asn1Element>) :
-    Asn1Structure(Tag.ASN1_SEQUENCE, children) {
+    Asn1Structure(Tag.SEQUENCE, children) {
 
     init {
         if (!tag.isConstructed) throw IllegalArgumentException("An ASN.1 Structure must have a CONSTRUCTED tag")
@@ -375,32 +515,36 @@ class Asn1Sequence internal constructor(children: List<Asn1Element>) :
  * ASN1 structure (i.e. containing child nodes) with custom tag
  */
 class Asn1CustomStructure private constructor(
-    tag: Tag, children: List<Asn1Element>
+    tag: Tag, children: List<Asn1Element>, sortChildren: Boolean
 ) :
-    Asn1Structure(tag, children) {
+    Asn1Structure(tag, children, sortChildren) {
     /**
      * ASN.1 CONSTRUCTED with custom tag
      * @param children the elements to put into this sequence
      * @param tag the custom tag to use
      * @param tagClass the tag class to use for this custom tag. defaults to [TagClass.UNIVERSAL]
+     * @param sortChildren whether to sort the passed child nodes. defaults to false
      */
     constructor(
         children: List<Asn1Element>,
         tag: ULong,
-        tagClass: TagClass = TagClass.UNIVERSAL
-    ) : this(Tag(tag, constructed = true, tagClass), children)
+        tagClass: TagClass = TagClass.UNIVERSAL,
+        sortChildren: Boolean = false
+    ) : this(Tag(tag, constructed = true, tagClass), children, sortChildren)
 
     /**
      * ASN.1 CONSTRUCTED with custom tag
      * @param children the elements to put into this sequence
      * @param tag the custom tag to use
      * @param tagClass the tag class to use for this custom tag. defaults to [TagClass.UNIVERSAL]
+     * @param sortChildren whether to sort the passed child nodes. defaults to false
      */
     constructor(
         children: List<Asn1Element>,
         tag: UByte,
-        tagClass: TagClass = TagClass.UNIVERSAL
-    ) : this(children, tag.toULong(), tagClass)
+        tagClass: TagClass = TagClass.UNIVERSAL,
+        sortChildren: Boolean = false
+    ) : this(children, tag.toULong(), tagClass, sortChildren)
 
 
     override val content: ByteArray by lazy {
@@ -420,9 +564,15 @@ class Asn1CustomStructure private constructor(
          * @param children the elements to put into this sequence
          * @param tag the custom tag to use
          * @param tagClass the tag class to use for this custom tag. defaults to [TagClass.UNIVERSAL]
+         * @param sort whether to sort the passed childr nodes. defaults to false
          */
-        fun asPrimitive(children: List<Asn1Element>, tag: ULong, tagClass: TagClass = TagClass.UNIVERSAL) =
-            Asn1CustomStructure(Tag(tag, constructed = false, tagClass), children)
+        fun asPrimitive(
+            children: List<Asn1Element>,
+            tag: ULong,
+            tagClass: TagClass = TagClass.UNIVERSAL,
+            sort: Boolean = false
+        ) =
+            Asn1CustomStructure(Tag(tag, constructed = false, tagClass), children, sort)
     }
 }
 
@@ -468,13 +618,13 @@ class Asn1PrimitiveOctetString(content: ByteArray) : Asn1Primitive(Tag.OCTET_STR
 /**
  * ASN.1 SET 0x31 ([BERTags.SET] OR [BERTags.CONSTRUCTED])
  */
-open class Asn1Set private constructor(children: List<Asn1Element>?, dontSort: Boolean) :
-    Asn1Structure(Tag.SET, if (dontSort) children else children?.sortedBy { it.tag }) {
+open class Asn1Set private constructor(children: List<Asn1Element>, dontSort: Boolean) :
+    Asn1Structure(Tag.SET, children, !dontSort) {
 
     /**
      * @param children the elements to put into this set. will be automatically sorted by tag
      */
-    internal constructor(children: List<Asn1Element>?) : this(children, false)
+    internal constructor(children: List<Asn1Element>) : this(children, false)
 
     init {
         if (!tag.isConstructed) throw IllegalArgumentException("An ASN.1 Structure must have a CONSTRUCTED tag")
@@ -500,10 +650,9 @@ open class Asn1Set private constructor(children: List<Asn1Element>?, dontSort: B
  * @param children the elements to put into this set. will be automatically checked to have the same tag and sorted by value
  * @throws Asn1Exception if children are using different tags
  */
-class Asn1SetOf @Throws(Asn1Exception::class) internal constructor(children: List<Asn1Element>?) :
-    Asn1Set(children?.let {
+class Asn1SetOf @Throws(Asn1Exception::class) internal constructor(children: List<Asn1Element>) :
+    Asn1Set(children.also { it ->
         if (it.any { elem -> elem.tag != it.first().tag }) throw Asn1Exception("SET OF must only contain elements of the same tag")
-        it.sortedBy { it.tag.encodedTag.encodeToString(Base16) } //TODO this is inefficient
     })
 
 /**