+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/pki/Pkcs10CertificationRequest.kt
@@ -1,10 +1,10 @@
 package at.asitplus.signum.indispensable.pki
 
-import at.asitplus.signum.indispensable.X509SignatureAlgorithm
 import at.asitplus.signum.indispensable.CryptoPublicKey
+import at.asitplus.signum.indispensable.X509SignatureAlgorithm
 import at.asitplus.signum.indispensable.asn1.*
 import at.asitplus.signum.indispensable.asn1.Asn1.BitString
-import at.asitplus.signum.indispensable.asn1.Asn1.Tagged
+import at.asitplus.signum.indispensable.asn1.Asn1.ExplicitlyTagged
 import at.asitplus.signum.indispensable.io.ByteArrayBase64Serializer
 import kotlinx.serialization.Serializable
 
@@ -50,8 +50,9 @@ data class TbsCertificationRequest(
 
         //subject Public Key
         +publicKey
-        +Tagged(0u) { attributes?.map { +it } }
+        +ExplicitlyTagged(0u) { attributes?.map { +it } }
     }
+
 
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
@@ -77,14 +78,14 @@ data class TbsCertificationRequest(
 
     companion object : Asn1Decodable<Asn1Sequence, TbsCertificationRequest> {
         @Throws(Asn1Exception::class)
-        override fun decodeFromTlv(src: Asn1Sequence) = runRethrowing {
+        override fun doDecode(src: Asn1Sequence) = runRethrowing {
             val version = (src.nextChild() as Asn1Primitive).readInt()
             val subject = (src.nextChild() as Asn1Sequence).children.map {
                 RelativeDistinguishedName.decodeFromTlv(it as Asn1Set)
             }
             val cryptoPublicKey = CryptoPublicKey.decodeFromTlv(src.nextChild() as Asn1Sequence)
             val attributes = if (src.hasMoreChildren()) {
-                (src.nextChild() as Asn1Tagged).verifyTag(0u)
+                (src.nextChild() as Asn1ExplicitlyTagged).verifyTag(0u)
                     .map { Pkcs10CertificationRequestAttribute.decodeFromTlv(it as Asn1Sequence) }
             } else null
 
@@ -143,7 +144,7 @@ data class Pkcs10CertificationRequest(
     companion object : Asn1Decodable<Asn1Sequence, Pkcs10CertificationRequest> {
 
         @Throws(Asn1Exception::class)
-        override fun decodeFromTlv(src: Asn1Sequence): Pkcs10CertificationRequest = runRethrowing {
+        override fun doDecode(src: Asn1Sequence): Pkcs10CertificationRequest = runRethrowing {
             val tbsCsr = TbsCertificationRequest.decodeFromTlv(src.nextChild() as Asn1Sequence)
             val sigAlg = X509SignatureAlgorithm.decodeFromTlv(src.nextChild() as Asn1Sequence)
             val signature = (src.nextChild() as Asn1Primitive).readBitString()