+++ indispensable/src/jvmTest/kotlin/at/asitplus/signum/indispensable/CastingTest.kt
@@ -11,7 +11,7 @@ class CastingTest : FreeSpec({
 
     "Primitive" {
         shouldThrow<Asn1StructuralException> { Asn1.Int(0).asSet() }
-        shouldThrow<Asn1StructuralException> { Asn1.Int(0).asTagged() }
+        shouldThrow<Asn1StructuralException> { Asn1.Int(0).asExplicitlyTagged() }
         shouldThrow<Asn1StructuralException> { Asn1.Int(0).asStructure() }
         shouldThrow<Asn1StructuralException> { Asn1.Int(0).asSequence() }
         shouldThrow<Asn1StructuralException> { Asn1.Int(0).asPrimitiveOctetString() }
@@ -21,7 +21,7 @@ class CastingTest : FreeSpec({
 
     "Set" {
         shouldThrow<Asn1StructuralException> { Asn1.Set { +Asn1.Null() }.asSequence() }
-        shouldThrow<Asn1StructuralException> { Asn1.Set { +Asn1.Null() }.asTagged() }
+        shouldThrow<Asn1StructuralException> { Asn1.Set { +Asn1.Null() }.asExplicitlyTagged() }
         shouldThrow<Asn1StructuralException> { Asn1.Set { +Asn1.Null() }.asPrimitiveOctetString() }
         shouldThrow<Asn1StructuralException> { Asn1.Set { +Asn1.Null() }.asEncapsulatingOctetString() }
 
@@ -31,7 +31,7 @@ class CastingTest : FreeSpec({
 
     "OctetString ENCAPSULATING" {
         shouldThrow<Asn1StructuralException> { Asn1.OctetStringEncapsulating { +Asn1.Null() }.asSet() }
-        shouldThrow<Asn1StructuralException> { Asn1.OctetStringEncapsulating { +Asn1.Null() }.asTagged() }
+        shouldThrow<Asn1StructuralException> { Asn1.OctetStringEncapsulating { +Asn1.Null() }.asExplicitlyTagged() }
         shouldThrow<Asn1StructuralException> { Asn1.OctetStringEncapsulating { +Asn1.Null() }.asPrimitiveOctetString() }
         shouldThrow<Asn1StructuralException> { Asn1.OctetStringEncapsulating { +Asn1.Null() }.asSequence() }
 
@@ -44,7 +44,7 @@ class CastingTest : FreeSpec({
         shouldThrow<Asn1StructuralException> { Asn1PrimitiveOctetString(byteArrayOf()).asSet() }
         shouldThrow<Asn1StructuralException> { Asn1PrimitiveOctetString(byteArrayOf()).asStructure() }
         shouldThrow<Asn1StructuralException> { Asn1PrimitiveOctetString(byteArrayOf()).asSequence() }
-        shouldThrow<Asn1StructuralException> { Asn1PrimitiveOctetString(byteArrayOf()).asTagged() }
+        shouldThrow<Asn1StructuralException> { Asn1PrimitiveOctetString(byteArrayOf()).asExplicitlyTagged() }
         shouldThrow<Asn1StructuralException> { Asn1PrimitiveOctetString(byteArrayOf()).asEncapsulatingOctetString() }
         Asn1PrimitiveOctetString(byteArrayOf()).let { it.asPrimitiveOctetString() shouldBe it }
         Asn1PrimitiveOctetString(byteArrayOf()).let { it.asPrimitive() shouldBe it }
@@ -52,13 +52,13 @@ class CastingTest : FreeSpec({
 
 
     "Custom Structure" {
-        shouldThrow<Asn1StructuralException> { Asn1.Tagged(19u) { +Asn1.Null() }.asSequence() }
-        shouldThrow<Asn1StructuralException> { Asn1.Tagged(19u) { +Asn1.Null() }.asSet() }
-        shouldThrow<Asn1StructuralException> { Asn1.Tagged(19u) { +Asn1.Null() }.asPrimitiveOctetString() }
-        shouldThrow<Asn1StructuralException> { Asn1.Tagged(19u) { +Asn1.Null() }.asEncapsulatingOctetString() }
+        shouldThrow<Asn1StructuralException> { Asn1.ExplicitlyTagged(19u) { +Asn1.Null() }.asSequence() }
+        shouldThrow<Asn1StructuralException> { Asn1.ExplicitlyTagged(19u) { +Asn1.Null() }.asSet() }
+        shouldThrow<Asn1StructuralException> { Asn1.ExplicitlyTagged(19u) { +Asn1.Null() }.asPrimitiveOctetString() }
+        shouldThrow<Asn1StructuralException> { Asn1.ExplicitlyTagged(19u) { +Asn1.Null() }.asEncapsulatingOctetString() }
 
-        Asn1.Tagged(19u) { +Asn1.Null() }.let { it.asStructure() shouldBe it }
-        Asn1.Tagged(19u) { +Asn1.Null() }.let { it.asTagged() shouldBe it }
+        Asn1.ExplicitlyTagged(19u) { +Asn1.Null() }.let { it.asStructure() shouldBe it }
+        Asn1.ExplicitlyTagged(19u) { +Asn1.Null() }.let { it.asExplicitlyTagged() shouldBe it }
     }